==> 1_local.ipynb <==
This notebook provide you a way to obtain a csv file for later QSAR building

import pandas as pd

Read structure

df3 = pd.read_csv('structures.txt')

df3

Next the txt file will be save as a csv file, then import back again, to beautify the structure.

df3.to_csv('structure.csv')

df4=pd.read_csv('structure.csv')

df4

selection=['smiles']
df4[selection]

df4_selection = df4[selection]
df4_selection

Read bioactivity

df7 = pd.read_csv('bioactivity.txt')

df7

df7.to_csv('bioactivity.csv')

df8=pd.read_csv('bioactivity.csv')

df8

selection=['bioactivity']
df8[selection]

df8_selection = df8[selection]
df8_selection

Combine structure and bioactivity

df10 = pd.concat([df4_selection, df8_selection], axis=1)
df10

Classification of bioactivities

bioactivity_class = []
for i in df10.bioactivity:
  if float(i) >= 7:
    bioactivity_class.append("active")
  elif float(i) <= 6:
    bioactivity_class.append("inactive")
  else:
    bioactivity_class.append("intermediate")

Combine 

bioactivity_class = pd.Series(bioactivity_class, name='bioactivity_class')
df11 = pd.concat([df10, bioactivity_class], axis=1)
df11

Save file as csv file, here we kept index

df12 = df11[df11.bioactivity_class != 'intermediate']
df12

df12.to_csv('sars-cov-2-local_bioactivity_data_preprocessed.csv')
