==> 3_build_local.ipynb <==
# **Bioinformatics Project - Computational Drug Discovery [Part 3] Descriptor Calculation and Dataset Preparation(Modified Version towards unknown smiles predcition purpose)**

Chanin Nantasenamat

[*'Data Professor' YouTube channel*](http://youtube.com/dataprofessor)

Modified by quantaosun@gmail.com


In **Part 3**, we will be calculating molecular descriptors that are essentially quantitative description of the compounds in the dataset. Finally, we will be preparing this into a dataset for subsequent model prediction in notebook 5.

---

## **Download PaDEL-Descriptor**

! wget https://github.com/dataprofessor/bioinformatics/raw/master/padel.zip
! wget https://github.com/dataprofessor/bioinformatics/raw/master/padel.sh

! unzip padel.zip

## **Load unknown smiles data, called "example.txt"**

Download the curated ChEMBL bioactivity data that has been pre-processed from Parts 1 and 2 of this Bioinformatics Project series. Here we will be using the **bioactivity_data_3class_pIC50.csv** file that essentially contain the pIC50 values that we will be using for building a regression model.

import pandas as pd

# We will convert a raw txt file with two lines of smiles to a standard dataframe by a csv intermediate

df3 = pd.read_csv('/content/sars-cov-2-local_bioactivity_data_preprocessed.csv')

df3

Next the txt file will be save as a csv file, then import back again, to beautify the structure.

selection=['smiles']
df3[selection]

df4_selection = df3[selection]
df4_selection

df4_selection.to_csv('molecule.smi', sep= '\t', index=False, header=False)

! cat molecule.smi | head -5

! cat molecule.smi | wc -l

## **Calculate fingerprint descriptors, for the unknown smiles**


### **Calculate PaDEL descriptors**

! cat padel.sh

! bash padel.sh

! ls -l

## **Preparing the X matrices**

### **X data matrix**

df3_X = pd.read_csv('descriptors_output.csv')

df3_X

df3_X = df3_X.drop(columns=['Name'])
df3_X

 # Y variable

df3_Y = df3['bioactivity']
df3_Y

# Combine X and Y

dataset3 = pd.concat([df3_X,df3_Y], axis=1)
dataset3

# You may have noticed that we now have a very long bit of 882 descriptor set, but when we built the model earlier, there are much less descriptors. This mismatch will be addressed later.

dataset3.to_csv('sars-cov-2-local_bioactivity_data_2class_pIC50_pubchem_fp.csv', index=False)

# **Let's download the CSV file to your local computer for the Part 3B (Model Building).**
